Added database example.  histograms int32 (instead of default int64).

+++ code/feature_extraction.py
@@ -29,7 +29,7 @@ def component_histograms(image, transform=None):
     assert len(shape) == 3
     _, _, num_comp = shape
     return [np.histogram(transform(image[:, :, [idx]]).flatten(),
-                         bins=NUM_HIST_BINS, range=(0, 256))[0]
+                         bins=NUM_HIST_BINS, range=(0, 256))[0].astype(np.int32)
             for idx in xrange(0, num_comp)]
 
 def get_labels(label_file):
@@ -50,9 +50,9 @@ def extract(image):
     gray_image = color.rgb2gray(image)
     gray_hist, _ = np.histogram(no_black(gray_image).flatten(),
                                 NUM_HIST_BINS, range=(0, 1))
-    return {'gray_hist': gray_hist, 'red_hist': r, 'green_hist': g,
-            'blue_hist':b, 'hue_hist': h, 'saturation_hist': s,
-            'value_hist': v}
+    return {'gray_hist': gray_hist.astype(np.int32), 'red_hist': r,
+            'green_hist': g, 'blue_hist':b, 'hue_hist': h,
+            'saturation_hist': s, 'value_hist': v}
 
 def extract_from_dir(directory):
     """Extract features from every jpeg file in directory (recursive)"""
